Move SnapshotsTestHelper to top-level
